# src/main/java/com/demo/revenue/controllers/VehicleApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.demo.revenue.controllers;

import java.util.Optional;

import javax.validation.Valid;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import com.demo.revenue.model.Vehicle;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;

@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2021-07-25T02:12:24.601+10:00[Australia/Sydney]")

@Validated
@Api(value = "vehicle", tags = { "vehicle" })
public interface VehicleApi {

	default Optional<NativeWebRequest> getRequest() {
		return Optional.empty();
	}

	/**
	 * POST /vehicle : Add a vehicle information
	 *
	 * @param body Vehicle object to be saved (required)
	 * @return Success (status code 200) or Invalid input (status code 400) or
	 *         Server error (status code 500)
	 * @throws Exception
	 */
	@ApiOperation(value = "Add a vehicle information", nickname = "saveVehicle", notes = "", response = Object.class, tags = {
			"vehicle", })
	@ApiResponses(value = {
			@ApiResponse(code = 200, message = "Success", response = Object.class),
			@ApiResponse(code = 400, message = "Invalid input"),
			@ApiResponse(code = 500, message = "Server error") })
	@RequestMapping(value = "/vehicle", produces = { "application/json" }, consumes = {
			"application/json" }, method = RequestMethod.POST)
	default ResponseEntity<Object> saveVehicle(
			@ApiParam(value = "Vehicle object to be saved", required = true) @Valid @RequestBody Vehicle body)
			throws Exception {
		return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

	}

}