# src/main/java/com/demo/revenue/dto/Person.java
package com.demo.revenue.dto;

import lombok.AllArgsConstructor;
import lombok.Data;

@AllArgsConstructor
@Data
public class Person {

	private Long id;

	private String firstname;

	private String lastname;

}