# src/main/java/com/demo/revenue/entity/Vehicle.java
package com.demo.revenue.entity;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.Table;

import lombok.Data;

@Entity
@Table(name = "Vehicle")
@Data
public class Vehicle {

	@Id
	@GeneratedValue(strategy = GenerationType.IDENTITY)
	private Long id;

	@Column(name = "MODEL")
	private String model;

	@Column(name = "COLOR")
	private String color;

	@Column(name = "PLATE_NO")
	private String plateNo;

}